input: Move InputManager to input.rs
Create input.rs and move input-related code there.
